feat: add docker compose and improve selfhosting (#60)
