+++ .env.example
@@ -1,26 +1,39 @@
-RAILS_ENV=
-NODE_ENV=
-
-# Postgres
+# REQUIRED if you're self-hosting
+RAILS_ENV= # set to 'production' or 'development'
+DATABASE_NAME=
 DATABASE_USER=
 DATABASE_PASSWORD=
+ALLOW_REGISTRATION=true # setting this to false disallows new sign ups
+SECRET_KEY_BASE= # Generate key with `rails secret` or manually (alphanumeric string with 128 chars)
+MAPBOX_ACCESS_TOKEN= # Create a mapbox.com account and retrieve your access token
 
-# Redis
-REDIS_URL=
 
-# Mapbox
-MAPBOX_ACCESS_TOKEN=
+# OPTIONAL if you're self-hosting
+INIT_USER_EMAIL= # set this if you run with docker-compose and want to create an initial user
+INIT_USER_PASSWORD= # set this if you run with docker-compose and want to create an initial user
+DATABASE_URL= # required if you're running Mapzy without docker-compose in production
+RAILS_LOG_TO_STDOUT=true # if you want to log errors in server logs
+MAILPACE_TOKEN= # if you want to use mailpace.com to handle email sending (e.g. forgot password email)
 
-# OhMySMTP
-OHMYSMTP_TOKEN=
 
-# Stripe
+# OPTIONAL when you're developing
 STRIPE_ENDPOINT_SECRET=
 STRIPE_MAXI_PRICE_ID=
 STRIPE_MEDIUM_PRICE_ID=
 STRIPE_MINI_PRICE_ID=
 STRIPE_SECRET_KEY=
+FUGU_API_KEY= # if you want to use fugu.lol for tracking
+FUGU_URL= # if you want to use fugu.lol for tracking
+DATABASE_NAME_TEST= # only needed if you are running automated tests during development
+REDIS_URL= # redis is currently only used for Fugu tracking and sending some of the emails
+MAPZY_CLOUD= # setting to "true" enables additional logic that we require for offering the Mapzy Cloud version
 
-# Fugu
-FUGU_API_KEY=
-FUGU_URL=
+
+
+
+
+
+
+
+
+
