+++ .github/workflows/docker.yml
@@ -1,0 +1,45 @@
+name: Docker
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  docker:
+    if: github.repository == 'mapzy/mapzy'
+    runs-on: ubuntu-latest
+    steps:
+      -
+        name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      -
+        name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      -
+        name: Login to DockerHub
+        uses: docker/login-action@v1 
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - 
+        name: Set Versions
+        uses: actions/github-script@v4
+        id: set_version
+        with:
+          script: |
+            const tag = context.ref.substring(10)
+            const no_v = tag.replace('v', '')
+            const dash_index = no_v.lastIndexOf('-')
+            const no_dash = (dash_index > -1) ?  no_v.substring(0, dash_index) : no_v
+            core.setOutput('tag', tag)
+            core.setOutput('no-v', no_v)
+            core.setOutput('no-dash', no_dash)
+      -
+        name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          file: Dockerfile.prod
+          tags: |
+            ${{secrets.DOCKERHUB_USERNAME}}/mapzy:latest
+            ${{secrets.DOCKERHUB_USERNAME}}/mapzy:${{steps.set_version.outputs.no-dash}}
