+++ .github/workflows/specs.yml
@@ -50,3 +50,7 @@ jobs:
       - name: Run specs
         run: |
           bundle exec rspec spec --exclude-pattern "**/features/*_spec.rb" --color --format progress
+        env:
+          STRIPE_ENDPOINT_SECRET: ${{ secrets.STRIPE_ENDPOINT_SECRET }}
+          ALLOW_REGISTRATION: true
+          MAPZY_CLOUD: true
