+++ CONTRIBUTING.md
@@ -1,12 +1,31 @@
+## Contribution Guide
+
+## Contribution Rules
+
+This is an open-source project and we welcome contributions that improve the code, UX/UI and add new features. However, this doesn't mean that we will accept all contributions. The best way to contribute is to check the open issues or ask us (bonjour@mapzy.io) if there's something you can contribute to.
+If you want to contribute, fork this repo and create a pull request with your changes that we can review.
+
+If you have any questions, head over to our [GitHub Discussions](https://github.com/mapzy/mapzy/discussions) page.
+
 ## Development
 
-## Gitpod
+### Setup
 
-We recommend developing with the excellent cloud-based development environment provided by [Gitpod](gitpod.io/). To start your ready-to-code development enviroment just add `gitpod.io/#` in front of the GitHub URL for any branch, or [click here](https://gitpod.io/#https://github.com/mapzy/mapzy) to get started with the main branch.
+There are a few ways you can get started with Mapzy on your local machine.
 
-Run `foreman start -f Procfile.dev` to start the Rails server and the Webpack Dev Server (or run the processes in separate terminals manually).
+### Gitpod
 
-### Local
+We recommend developing with the excellent cloud-based development environment provided by [Gitpod](gitpod.io/). To start your ready-to-code development enviroment just add `gitpod.io/#` in front of the GitHub URL for any branch, or [click here](https://gitpod.io/#https://github.com/mapzy/mapzy) to get started with the main branch. Don't forget to add the environment variables from `.env.example` to your Gitpod account.
+
+### Docker
+
+You can build the Docker image using the `Dockerfile.dev` configuration. If you don't want to run Postgres or Redis on your machine, you can use our `docke-compose` file to start them via Docker. Just run `docker-compose -f docker-compose.dev.yml up`.
+
+Don't forget to create an `.env` file using `.env.example` as a template.
+
+If you're not using `docker-compose`, make sure to create the database with `docker run --env-file .env your_image rails db:create` before starting the Rails server.
+
+### Old school
 
 Dependencies:
 - ruby ~3.0.0
@@ -21,9 +40,13 @@ cp .env.example .env
 
 # Create and migrate the database
 rails db:create db:migrate
-
-# Finally, run the app locally
-rails s -b 0.0.0.0
 ```
 
-Run `foreman start -f Procfile.dev` to start the Rails server and the Webpack Dev Server (or run the processes in separate terminals manually).
+## Starting the development server
+To start the server, run `.foreman start -f Procfile.dev`. This starts the Rails server, the Tailwind watch process and the Sidekiq process.
+
+
+## Mapzy Cloud
+There are a few places in the codebase where you will encounter logic that is only needed for Mapzy Cloud, the hosted solution by us. These parts of the code are not needed if you are self-hosting, and probably also not relevant for you as a contributor. For the most part this concerns code that relates to payment and subscriptions.
+
+Currently, we use the environment variable `MAPZY_CLOUD` to determine if the instance is being self-hosted or not.
