+++ Dockerfile.dev
@@ -19,7 +19,6 @@ ENV BUNDLE_CACHE_ALL true
 ENV APP_HOME /app
 ENV RAILS_ENV development
 ENV RACK_ENV development
-ENV NODE_ENV development
 
 # Set working directory
 WORKDIR $APP_HOME
@@ -34,18 +33,19 @@ RUN bundle config --global frozen 1 && \
   find /usr/local/bundle/gems/ -name "*.o" -delete
 
 # Copy app files
-ADD . $APP_PATH
+ADD . .
 
 ###############################################################################
 # Stage 2: Run
 FROM ruby:3.0.0
 
-RUN mkdir -p /app
-WORKDIR /app
-
 ENV RAILS_ENV development
-ENV NODE_ENV development
 ENV APP_HOME /app
+
+RUN mkdir -p $APP_HOME
+WORKDIR $APP_HOME
+
+EXPOSE 3000
 
 # Copy necessary data at runtime
 COPY --from=builder /usr/lib /usr/lib
@@ -55,3 +55,6 @@ COPY --from=builder /usr/local/bundle /usr/local/bundle
 
 # Copy app files
 COPY --from=builder $APP_HOME $APP_HOME
+
+# Default command (overriden if used with this project's docker-compose)
+CMD ["rails", "server", "-b", "0.0.0.0"]
