+++ Dockerfile.prod
@@ -21,7 +21,6 @@ ENV BUNDLE_CACHE_ALL true
 ENV APP_HOME /app
 ENV RAILS_ENV production
 ENV RACK_ENV production
-ENV NODE_ENV production
 
 # Set working directory
 WORKDIR $APP_HOME
@@ -36,7 +35,7 @@ RUN bundle config --global frozen 1 && \
   find /usr/local/bundle/gems/ -name "*.o" -delete
 
 # Copy app files
-ADD . $APP_PATH
+COPY . .
 
 # Precompile assets
 # SECRET_KEY_BASE=`bin/rake secret` is added here as a workaround for
@@ -48,13 +47,14 @@ RUN SECRET_KEY_BASE=`bin/rake secret` rails assets:precompile --trace && \
 # Stage 2: Run
 FROM ruby:3.0.0
 
-RUN mkdir -p /app
-WORKDIR /app
-
 ENV RAILS_ENV production
-ENV NODE_ENV production
 ENV RAILS_SERVE_STATIC_FILES true
 ENV APP_HOME /app
+
+RUN mkdir -p $APP_HOME
+WORKDIR $APP_HOME
+
+EXPOSE 3000
 
 # Copy necessary data at runtime
 COPY --from=builder /usr/lib /usr/lib
@@ -64,3 +64,9 @@ COPY --from=builder /usr/local/bundle /usr/local/bundle
 
 # Copy app files
 COPY --from=builder $APP_HOME $APP_HOME
+
+# Default entrypoint (overriden if used with this project's docker-compose)
+ENTRYPOINT ["sh", "/app/docker-entrypoint.sh"]
+
+# Default command (overriden if used with this project's docker-compose)
+CMD ["rails", "server", "-b", "0.0.0.0"]
