+++ Gemfile
@@ -6,7 +6,7 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 ruby "3.0.0"
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
-gem "rails", "~> 6.1.4", ">= 6.1.4.4"
+gem "rails", "~> 6.1.4"
 # Use postgresql as the database for Active Record
 gem "pg", "~> 1.1"
 # Use Puma as the app server
@@ -47,7 +47,7 @@ gem "stripe", "~> 5.41"
 
 gem "foreman", "~> 0.87.2"
 
-gem "ohmysmtp-rails", "~> 0.1.9"
+gem "mailpace-rails", "~> 0.3.0"
 
 gem "sidekiq", "~> 6.4"
