+++ Procfile.docker.dev
@@ -1,0 +1,3 @@
+web: bundle exec rails server -p 3000 -b '0.0.0.0'
+tailwind: rails tailwindcss:watch
+sidekiq: bundle exec sidekiq -q default -q mailers
