+++ app/models/user.rb
@@ -20,7 +20,6 @@
 #
 class User < ApplicationRecord
   include Hashid::Rails
-  include Sidekiq::Worker
 
   has_many :maps, dependent: :destroy
   has_one :account, dependent: :destroy
@@ -51,16 +50,5 @@ class User < ApplicationRecord
 
   def create_account
     Account.create(user: self)
-  end
-
-  def setup_email_workers
-    EmailWorker.perform_at(7.days.from_now, "reminder_email1", id)
-    EmailWorker.perform_at(13.days.from_now, "reminder_email2", id)
-    EmailWorker.perform_at(14.days.from_now, "account_inactivated_email", id)
-    AccountWorker.perform_at(14.days.from_now, id)
-  end
-
-  def send_welcome_email
-    AccountMailer.with(email: email).welcome_email.deliver_later
   end
 end
