+++ config/initializers/sidekiq.rb
@@ -2,12 +2,12 @@
 
 Sidekiq.configure_server do |config|
   config.redis = {
-    url: ENV['REDIS_URL'] || 'redis://localhost:6379'
+    url: ENV["REDIS_URL"] || "redis://localhost:6379"
   }
 end
 
 Sidekiq.configure_client do |config|
   config.redis = {
-    url: ENV['REDIS_URL'] || 'redis://localhost:6379'
+    url: ENV["REDIS_URL"] || "redis://localhost:6379"
   }
 end
