+++ config/initializers/stripe.rb
@@ -1,1 +1,1 @@
-Stripe.api_key = ENV['STRIPE_SECRET_KEY']
+Stripe.api_key = ENV["STRIPE_SECRET_KEY"]
